Adding a recursive option (default: on)
If the user wants to have the diff not be recurive, the global value
'g:DirDiffRecursive' can be set to '0'.  This controls whether '-r' will
be on the `diff` command line.

It works like the existing options, if not set, it is set by this plugin
to its default value, '1'.